UnfoldDecode

Documentation for UnfoldDecode.

This toolbox right now implements:

- Overlapcorrected Decoding ala [Gal Vishne, Leon Deouell et al.](https://doi.org/10.1101/2023.06.28.546397)
- B2B regression (Jean-Remy King et al. ) - still in Unfold.jl but we should convert soonish :)

StatsAPI.coeftable

StatsAPI.fit

UnfoldDecode.fit_timepoints

UnfoldDecode.singletrials

StatsAPI.coeftable — Method

measure: Any measure of MLJ. Find a whole list: MLJ.measures(), by default balancedaccuracy, but rootmean_squared, accuracy etc. all possible.

StatsAPI.fit — Method

using Unfold: @maybe_threads Use UnfoldLinearModelContinuous time to remove all overlap from an event. Use the model classifier to predict the EEG data. Everything runs cross-validated with nfolds

model: By default LDA(), but could be any MLJ machine with specified parameters

target: String or Symbol with the tbl[:,column] to be decoded

UnfoldFitkwargs: optional Named Tuple as kwargs to provide to the initial "overlap-cleaning" modelfit, e.g. UnfoldFitkwargs = (;solver=(x,y)->solver_krylov(x,y,GPU=true)) for GPU fit (need to load Krylov and CUDA before)

UnfoldDecode.fit_timepoints — Method

For each time-point, fit a MLJ-Machine

UnfoldDecode.singletrials — Method

Returns the single trial corrected (and "epoched") trials from the basis target